requirements.txt
Flask==2.3.2
python-telegram-bot==20.3
python-dotenv==1.0.0
xrpl-py==1.7.2
